packages/driver/pyproject.toml
[project]
name = "displaypad-driver"
version = "0.2.0"
description = "A Driver to communicate with the mountain displaypad."
authors = [
    {name = "AnnikenToGo",email = "anniken@annikentogo.de"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "pillow (>=12.1.0,<13.0.0)",
    "pyusb (>=1.3.1,<2.0.0)",
    "libusb (>=1.0.29.post1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
